+++ src/webview/index.ts
@@ -1,2 +1,2 @@
-export { TimelineView } from './timeline/view';
-export { NotificationsView } from './notifications/view';
+export { TimelineView } from './pages/timeline';
+export { NotificationsView } from './pages/notifications';
